fix trainer

+++ scripts/main.py
@@ -3,13 +3,19 @@ import json
 import os
 
 from modules.trainer import Trainer
+from utils.argument_handler import argment_handler
 
 if __name__ == "__main__":
     config = json.load(open('config.json', 'r'))
     config["output_path"] += "{:%Y-%m-%d_%H:%M}/".format(datetime.datetime.now())
 
-    os.mkdir(config["output_path"])
-    json.dump(config, open(config["output_path"] + 'config.json', 'w'), indent=4)
-    
+    args = argment_handler()
+    if args.no_write:
+        config['is_file_saved'] = not args.no_write
+
+    if config['is_file_saved']:
+        os.mkdir(config["output_path"])
+        json.dump(config, open(config["output_path"] + 'config.json', 'w'), indent=4)
+
     t = Trainer(**config)
     t.optimize()
